+++ pypi/pyoz/__main__.py
@@ -1,0 +1,5 @@
+"""Allow running pyoz as: python -m pyoz"""
+
+from pyoz import main
+
+main()